pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "psycopg2-binary>=2.9.11",
    "pycountry>=24.6.1",
    "python-whois>=0.9.6",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "scapy>=2.6.1",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]
